Доработка дз 5. Вроде получилось
